# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-fibonacci",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the `n`th Fibonacci number (0-indexed).",
        "description": "This service owns the *control flow* — an iterative fold over the sequence —\nbut does no arithmetic of its own: each successive term is produced by\nasking [`srvcs-add`] for `a + b`. `fibonacci(0) == 0`, `fibonacci(1) == 1`,\n`fibonacci(10) == 55`.\n\nA negative index is rejected with `422` locally. If `srvcs-add` is\nunreachable it reports itself degraded (`503`); if `srvcs-add` rejects an\noperand it forwards the `422`; and if the loop somehow exceeds its iteration\ncap it returns `500` rather than spinning.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FibonacciResponse"
                }
              }
            }
          },
          "422": {
            "description": "n is negative, or a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "the fibonacci loop exceeded its iteration cap"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FibonacciResponse": {
        "type": "object",
        "required": [
          "value",
          "result"
        ],
        "properties": {
          "result": {
            "type": "integer",
            "format": "int64"
          },
          "value": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}